+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added `Kord#getInvite`. #143
 * Added pending property to member. #153
 * Added missing contracts to various builders. #243
+* Added message reference core entity. #245
 
 ## Fixes